requirements.txt
gradio>=4.0.0
Pillow>=9.5.0
pandas>=2.1.0
numpy>=1.24.0
jiwer>=3.0.0
tensorboard>=2.14.0
opencv-python-headless>=4.8.0
tqdm>=4.66.0
transformers>=4.36.0
